pyproject.toml
[tool.poetry]
name = "llm-pipeline-system"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.extras]
training = ["unsloth"]

[tool.poetry.dependencies]
python = "~3.11"
pymongo = "^4.6.2"
click = "^8.0.1"
loguru = "^0.7.2"
rich = "^13.7.1"
numpy = "^1.26.4"
poethepoet = "0.29.0"
datasets = "^3.0.1"
torch = "2.2.2"

# Digital data ETL
selenium = "^4.21.0"
webdriver-manager = "^4.0.1"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"
jmespath = "^1.0.1"
chromedriver-autoinstaller = "^0.6.4"

# Feature engineering
qdrant-client = "^1.8.0"
langchain = "^0.2.11"
sentence-transformers = "^3.0.0"

# RAG
langchain-openai = "^0.1.3"
langchain-google-genai = "^1.0.0"
jinja2 = "^3.1.4"
tiktoken = "^0.7.0"
fake-useragent = "^1.5.1"
langchain-community = "^0.2.11"

# Inference (Local alternatives to AWS SageMaker)
fastapi = ">=0.100,<=0.110"
uvicorn = "^0.30.6"
bentoml = ">=1.2.0,<1.3.0"
transformers = ">=4.31.0"
huggingface-hub = ">=0.23.0"
gradio = "^4.0.0"

# Fix dependency conflicts
defusedxml = "*"
tinycss2 = "*"
markdown = ">=2.6.8"
h5py = ">=2.9.0"
protobuf = "<5.0.0,>=3.20.3"
pymysql = "^1.1.0"
wrapt = "<1.15,>=1.11.0"
accelerate = ">=0.34.0"
clearml = "^1.9.0"
trl = "^0.7.0"
mlflow = "^2.14.3"

# Training dependencies (optional)
unsloth = {version = "*", optional = true}




[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
pre-commit = "^3.7.1"
pytest = "^8.2.2"


[tool.poetry.group.aws.dependencies]
sagemaker = ">=2.232.2"
s3fs = ">2022.3.0"
aws-profile-manager = "^0.7.3"
kubernetes = "^30.1.0"
sagemaker-huggingface-inference-toolkit = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ----------------------------------
# --- Poe the Poet Configuration ---
# ----------------------------------

[tool.poe.tasks]
# Data pipelines
run-digital-data-etl = "poetry run python -m tools.run --run-etl --no-cache --etl-config-filename digital_data_etl.yaml"
run-feature-engineering-pipeline = "poetry run python -m tools.run --no-cache --run-feature-engineering"
run-generate-instruct-datasets-pipeline = "poetry run python -m tools.run --no-cache --run-generate-instruct-datasets"
run-generate-preference-datasets-pipeline = "poetry run python -m tools.run --no-cache --run-generate-preference-datasets"
run-end-to-end-data-pipeline = "poetry run python -m tools.run --no-cache --run-end-to-end-data"

# Utility pipelines
run-export-artifact-to-json-pipeline = "poetry run python -m tools.run --no-cache --run-export-artifact-to-json"
run-export-data-warehouse-to-json = "poetry run python -m tools.data_warehouse --export-raw-data"
run-import-data-warehouse-from-json = "poetry run python -m tools.data_warehouse --import-raw-data"

# Training pipelines
run-training-pipeline = "poetry run python -m tools.run --no-cache --run-training"
run-evaluation-pipeline = "poetry run python -m tools.run --no-cache --run-evaluation"

# Inference
call-rag-retrieval-module = "poetry run python -m tools.rag"

run-inference-ml-service = "poetry run uvicorn tools.ml_service:app --host 0.0.0.0 --port 8000 --reload"
call-inference-ml-service = "curl -X POST 'http://127.0.0.1:8000/rag' -H 'Content-Type: application/json' -d '{\"query\": \"My name is User. Could you draft a LinkedIn post discussing RAG systems? I am particularly interested in how RAG works and how it is integrated with vector DBs and LLMs.\"}'"

## Local Deployment (Free alternatives to AWS SageMaker)
deploy-local = "poetry run python -m tools.deploy_local deploy-local"
deploy-hf-spaces = "poetry run python -m tools.deploy_local deploy-hf-spaces"
deployment-status = "poetry run python -m tools.deploy_local status"
test-local-endpoint = "poetry run python -m tools.deploy_local test-inference --endpoint-url http://localhost:8000"
delete-deployment = "poetry run python -m tools.deploy_local delete"
list-deployments = "poetry run python -m tools.deploy_local list-deployments"

## Legacy AWS (Deprecated)
create-sagemaker-role = "echo 'AWS SageMaker is deprecated. Use: poetry poe deploy-local'"
create-sagemaker-execution-role = "echo 'AWS SageMaker is deprecated. Use: poetry poe deploy-local'"
deploy-inference-endpoint = "echo 'AWS SageMaker is deprecated. Use: poetry poe deploy-local'"
test-sagemaker-endpoint = "echo 'AWS SageMaker is deprecated. Use: poetry poe test-local-endpoint'"
delete-inference-endpoint = "echo 'AWS SageMaker is deprecated. Use: poetry poe delete-deployment'"

## Docker
build-docker-image = "docker buildx build --platform linux/amd64 -t llmtwin -f Dockerfile ."
run-docker-end-to-end-data-pipeline = "docker run --rm --network host --shm-size=2g --env-file .env llmtwin poetry poe --no-cache --run-end-to-end-data"
bash-docker-container = "docker run --rm -it --network host --env-file .env llmtwin bash"

# QA
lint-check = "poetry run ruff check ."
format-check = "poetry run ruff format --check ."
lint-check-docker = "sh -c 'docker run --rm -i hadolint/hadolint < Dockerfile'"
gitleaks-check = "docker run -v .:/src zricethezav/gitleaks:latest dir /src/llm_pipeline_system"
lint-fix = "poetry run ruff check --fix ."
format-fix = "poetry run ruff format ."

# Tests
[tool.poe.tasks.test]
cmd = "poetry run pytest tests/"
env = { ENV_FILE = ".env.testing" }

[tool.poe.tasks.chat]
cmd = "poetry run python -m tools.chat_test"
help = "Run the chat interface to test the RAG system."
